별점 총계 ajax 반영 완료 / 전체 목록 조회 시에도 보일 수 있게 xml 쿼리 변경

+++ ohTravel/src/main/java/com/oracle/ohTravel/review/dao/ReviewDAO.java
@@ -13,6 +13,6 @@ public interface ReviewDAO {
 	
 	int reviewCnt(String rv_real_id) throws Exception;
 
-	int selectAvgRating(ReviewDTO reviewDTO);
+	double selectAvgRating(ReviewDTO reviewDTO);
 
 }
